Remove GET_ACCOUNTS permission (and require M+).
Since we only access accounts managed by our own account manager, we
don't actually need GET_ACCOUNTS as of Android M+. We could continue
to declare GET_ACCOUNTS with maxSdkVersion="M" to keep older platforms
working, but this has policy requirements which come along with it
that are not worth the trouble.

Also get the build working with the latest Android Studio and bump the
version in preparation for a release.